requirements.txt
django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
mysqlclient==2.2.7
pillow==11.2.1
python-dotenv==1.0.1 